Fixed and finished DTO, createAccount, getAccountById and deleteAccountById methods in controller and service.

+++ src/main/java/com/example/banksystemapi/controller/AccountController.java
@@ -1,11 +1,16 @@
 package com.example.banksystemapi.controller;
 
 import com.example.banksystemapi.dto.AccountDto;
+import com.example.banksystemapi.model.Account;
+import com.example.banksystemapi.repository.AccountRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.example.banksystemapi.services.AccountService;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/accounts")
@@ -14,16 +19,19 @@ public class AccountController {
 
     private final AccountService accountService;
 
+    private final AccountRepository accountRepository;
+
+
     @PostMapping("/")
-    public ResponseEntity<AccountDto> createAccount(@RequestBody AccountDto request) {
-        String name = request.getName();
-        double amount = request.getAmount();
+    public ResponseEntity<AccountDto> createAccount(@RequestBody AccountDto accountDto) throws Exception {
+        try {
+            AccountDto createdAccount = accountService.createAccount(accountDto);
+            return ResponseEntity.ok(createdAccount);
+        } catch (Exception e) {
 
-        AccountDto createdAccount = accountService.createAccount(name, amount);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdAccount);
+           throw  new Exception("Failed to create account: "+ e.getMessage());
+        }
     }
-
-
 
     @GetMapping("/{id}")
     public ResponseEntity<AccountDto> getAccountById(@PathVariable("id") Integer id) {
@@ -36,8 +44,20 @@ public class AccountController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteAccountById(@PathVariable("id") Integer id) {
-        accountService.deleteAccountById(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deleteAccountById(@PathVariable("id") Integer id) {
+        try {
+            Optional<Account> optionalAccount = accountRepository.findById(id);
+            if (optionalAccount.isPresent()) {
+                accountService.deleteAccountById(id);
+                return ResponseEntity.ok("Account with ID " + id + " deleted successfully");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Account with ID " + id + " does not exist");
+            }
+        } catch (Exception e) {
+            // Log the error or handle it appropriately
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to delete account: " + e.getMessage());
+        }
     }
+
+
 }
